+++ src/Observable.spec.ts
@@ -1,4 +1,4 @@
-import { Observable } from './Observable'
+import { Observable, NO_VALUE } from './Observable'
 import { DispatcherPriorities } from './DispatcherPriorities'
 import { dispatch, flush } from './DispatcherInstance'
 const { ALL } = DispatcherPriorities
@@ -70,6 +70,12 @@ describe('Observable', () => {
     it('can release a promise of next value', async () => {
         const val = await Observable.of(1).toPromise()
         expect(val).toBe(1)
+    })
+
+    it('can create stream from a promise', async () => {
+        expect(
+            await Observable.fromPromise(Promise.resolve(6)).toPromise()
+        ).toBe(6)
     })
 
     it('can filter messages', done => {
@@ -378,14 +384,14 @@ describe('Observable', () => {
     it('should reemit last value', async () => {
         const a = new Observable().startWith(1)
         const a2 = a.reemit()
-        expect(a.previous()).toBeUndefined()
-        expect(a2.previous()).toBeUndefined()
+        expect(a.previous()).toBe(NO_VALUE)
+        expect(a2.previous()).toBe(NO_VALUE)
 
         expect(await a.toPromise()).toBe(1)
         expect(await a2.toPromise()).toBe(1)
 
         const b = a.reemit()
-        expect(b.previous()).toBeUndefined()
+        expect(b.previous()).toBe(NO_VALUE)
         expect(await b.toPromise()).toBe(1)
     })
 })